+++ probe/endpoint/connection_tracker.go
@@ -1,3 +1,5 @@
+// +build linux
+
 package endpoint
 
 import (
