+++ probe/endpoint/nat.go
@@ -1,3 +1,5 @@
+// +build linux
+
 package endpoint
 
 import (
