+++ probe/endpoint/nat_internal_test.go
@@ -1,3 +1,5 @@
+// +build linux
+
 package endpoint
 
 import (
